+++ src/snowflake/cli/app/commands_registration/builtin_plugins.py
@@ -3,6 +3,11 @@ from snowflake.cli.plugins.git import plugin_spec as git_plugin_spec
 from snowflake.cli.plugins.nativeapp import plugin_spec as nativeapp_plugin_spec
 from snowflake.cli.plugins.notebook import plugin_spec as notebook_plugin_spec
 from snowflake.cli.plugins.object import plugin_spec as object_plugin_spec
+
+# TODO 3.0: remove this import
+from snowflake.cli.plugins.object_stage_deprecated import (
+    plugin_spec as object_stage_deprecated_plugin_spec,
+)
 from snowflake.cli.plugins.snowpark import plugin_spec as snowpark_plugin_spec
 from snowflake.cli.plugins.spcs import plugin_spec as spcs_plugin_spec
 from snowflake.cli.plugins.sql import plugin_spec as sql_plugin_spec
@@ -23,6 +28,7 @@ def get_builtin_plugin_name_to_plugin_spec():
         "streamlit": streamlit_plugin_spec,
         "git": git_plugin_spec,
         "notebook": notebook_plugin_spec,
+        "object-stage-deprecated": object_stage_deprecated_plugin_spec,
     }
 
     return plugin_specs